% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{print.sfcqr}
\alias{print.sfcqr}
\title{Print a supervised functional censored quantile fit}
\usage{
\method{print}{sfcqr}(x, ...)
}
\arguments{
\item{x}{A fitted \code{\link[=sfcqr]{sfcqr()}} object.}

\item{...}{Reserved for future extensions.}
}
\value{
The fit, invisibly.
}
\description{
Print a supervised functional censored quantile fit
}
